feat: add AvailableLeaveDays column and LeavePolicies table

+++ src/Application/Ophthalmologists/Common/OphthalmologistDto.cs
@@ -19,7 +19,7 @@ public class OphthalmologistDto
     public string? DegreeUrl { get; set; }
     public decimal RatingAverage { get; set; }
     public int RatingCount { get; set; }
-    public decimal AvailableLeaveDays { get; set; }
+    public int AvailableLeaveDays { get; set; }
     public decimal ConsultationFee { get; set; }
     public decimal? MinPrice { get; set; }
     public decimal? MaxPrice { get; set; }
