+++ src/Application/SystemAdmin/LeavePolicies/Commands/CreateLeavePolicy/CreateLeavePolicyCommand.cs
@@ -6,6 +6,6 @@ namespace Application.SystemAdmin.LeavePolicies.Commands.CreateLeavePolicy;
 public record CreateLeavePolicyCommand : ICommand<LeavePolicyDto>
 {
     public string Name { get; init; } = string.Empty;
-    public decimal AdditionalDays { get; init; }
+    public int AdditionalDays { get; init; }
     public string? Description { get; init; }
 }
