+++ src/Application/SystemAdmin/LeavePolicies/Commands/UpdateLeavePolicy/UpdateLeavePolicyCommand.cs
@@ -6,6 +6,6 @@ public record UpdateLeavePolicyCommand : ICommand
 {
     public Guid PolicyId { get; init; }
     public string Name { get; init; } = string.Empty;
-    public decimal AdditionalDays { get; init; }
+    public int AdditionalDays { get; init; }
     public string? Description { get; init; }
 }
